iot_smart-agri/api/requirements.txt
fastapi
uvicorn
pymongo
paho-mqtt
pydantic
python-dotenv